+++ ReleaseNotes.md
@@ -28,3 +28,4 @@
 - Add Mile::WinRT::Property template class.
 - Add Mile::WinRT::Event template class.
 - Add Mile::WinRT::IsPackagedMode method.
+- Add MileIsCurrentProcessElevated function.